[update] 디테일 창 코드 리팩토링 완료
