+++ src/api/MyActivitesApi.js
@@ -1,4 +1,3 @@
-import axios from 'axios';
 import { jsonApi } from './apiInstance';
 
 // 즐겨찾기 추가나 리뷰를 작성한 가게들 불러오기
@@ -28,22 +27,14 @@ export const addMyActivity = async ({ queryKey, post }) => {
   await jsonApi.post(`/${queryKey}`, post);
 };
 
+// 즐겨찾기, 리뷰 삭제 api
+export const removeMyActivity = async ({ queryKey, id }) => {
+  await jsonApi.delete(`/${queryKey}/${id}`);
+};
+
 // 리뷰 수정 api
 export const updateMyActivity = async ({ queryKey, id, content }) => {
   await jsonApi.patch(`${queryKey}/${id}`, {
     review: content
   });
-};
-
-// 즐겨찾기 수정 api
-export const updateFavorite = async ({ id, content }) => {
-  const response = await jsonApi.patch(`/favorites/${id}`, {
-    favorite: content
-  });
-  return response.data;
-};
-
-// 즐겨찾기, 리뷰 삭제 api
-export const removeMyActivity = async ({ queryKey, id }) => {
-  await jsonApi.delete(`/${queryKey}/${id}`);
 };
